requirements.txt
aiosqlite==0.21.0
altair==5.0.1
annotated-types==0.7.0
ansi2html==1.8.0
anyio==3.7.1
appnope==0.1.3
astro==0.2.7
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
blinker==1.6.2
bokeh==2.4.3
branca==0.8.1
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
configparser==7.0.0
contourpy==1.1.0
cryptography==43.0.3
cssselect==1.2.0
cycler==0.11.0
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
docopt==0.6.2
duckdb==0.8.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
extra-streamlit-components==0.1.56
fake-useragent==1.5.1
Faker==37.1.0
favicon==0.7.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
Flask==3.0.0
Flask-Compress==1.14
folium==0.19.4
fonttools==4.41.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.10.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.23.0
google-auth-oauthlib==1.1.0
googletrans==4.0.0rc1
graphviz==0.20.3
gspread==5.11.1
gspread-dataframe==3.3.1
gspread-formatting==1.1.2
gspread-pandas==3.2.3
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
htbuilder==0.9.0
html5lib==1.1
httpcore==0.13.7
httpx==0.13.3
huggingface-hub==0.27.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.7.0
ipykernel==6.22.0
ipython==8.13.1
itsdangerous==2.1.2
jedi==0.18.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.4.2
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kinematics==0.1.7
kiwisolver==1.4.4
langchain-core==0.3.27
langchain-ollama==0.2.2
langsmith==0.2.4
lml==0.1.0
Logbook==1.7.0.post0
lxml==5.3.0
lxml_html_clean==0.4.1
Markdown==3.8
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
multitasking==0.0.11
nest-asyncio==1.5.6
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.25.0
oauthlib==3.2.2
ollama==0.4.4
opencv-python==4.10.0.84
openpyxl==3.1.2
orjson==3.10.12
outcome==1.2.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
peewee==3.17.8
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==24.0
pipwin==0.5.2
platformdirs==3.11.0
plotly==5.15.0
prometheus_client==0.21.1
prompt-toolkit==3.0.38
propcache==0.3.0
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pybind11==2.10.4
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.8.1b0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
Pygments==2.15.1
pyjsparser==2.7.1
PyJWT==2.7.0
pymdown-extensions==10.14.3
Pympler==1.0.1
PyMySQL==1.1.1
pyodbc==5.0.1
pyOpenSSL==25.0.0
pyparsing==3.0.9
pypdfium2==4.30.1
PyPrind==2.11.3
pypyodbc==1.3.6
pyrsistent==0.19.3
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
rank-bm25==0.2.2
regex==2024.11.6
Represent==2.1
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3986==1.5.0
rich==13.4.2
rsa==4.9
rush==2021.4.0
safetensors==0.4.5
SciPy==1.11.1
seaborn==0.13.2
selenium==4.11.2
setuptools==67.6.1
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0.post7
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
spacetrack==1.3.0
spiceypy==6.0.0
sql-metadata==2.8.0
SQLAlchemy==2.0.36
sqlparse==0.4.4
squarify==0.4.4
st-annotated-text==4.0.2
st-theme==1.2.3
stack-data==0.6.2
streamlit==1.41.1
streamlit-authenticator==0.2.2
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit_folium==0.24.0
streamlit-gsheets==0.0.1
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-option-menu==0.3.6
streamlit-player==0.1.5
streamlit-timeline==0.0.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.1
ta==0.11.0
TA-Lib==0.4.28
tenacity==8.2.2
tesseract==0.1.3
texttable==1.6.7
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.0
torch==2.5.1
torchvision==0.20.1
tornado==6.3.1
tqdm==4.67.1
traitlets==5.9.0
transformers==4.47.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.3.58
ultralytics-thop==2.0.13
urllib3==1.26.16
validators==0.20.0
virtualenv==20.24.6
vnai==0.1.3
wcwidth==0.2.6
webdriver-manager==4.0.2
xlrd==2.0.1
xlwt==1.3.0
xxhash==3.5.0
streamlit-option-menu 0.4.0
